# Documentazione/OpenAPI/openapi.yaml
openapi: 3.0.3
info:
  title: BugBoard26 API
  version: 1.0.0
  description: API surface for BugBoard26.JWT session revocation remains enabled as
    an intentional security exception.
paths:
...
  /api/projects/{projectId}/members:
    get:
      operationId: projects_members_list
      parameters:
      - in: query
        name: includeAdmins
        schema:
          type: boolean
      - in: path
        name: projectId
        schema:
          type: integer
        description: A unique integer value identifying this project.
        required: true
      tags:
      - Projects
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectMembership'
          description: ''
  /api/projects/{projectId}/subscriptions/me:
    get:
      operationId: projects_subscriptions_me_retrieve
      description: Admin subscription state for the authenticated user on the project.
      summary: Project subscription
      parameters:
      - in: path
        name: projectId
        schema:
          type: integer
        description: A unique integer value identifying this project.
        required: true
      tags:
      - Projects
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSubscriptionState'
          description: ''
        '204':
          description: Subscription updated
    put:
      operationId: projects_subscriptions_me_update
      description: Admin subscription state for the authenticated user on the project.
      summary: Project subscription
      parameters:
      - in: path
        name: projectId
        schema:
          type: integer
        description: A unique integer value identifying this project.
        required: true
      tags:
      - Projects
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSubscriptionState'
          description: ''
        '204':
          description: Subscription updated
    delete:
      operationId: projects_subscriptions_me_destroy
      description: Admin subscription state for the authenticated user on the project.
      summary: Project subscription
      parameters:
      - in: path
        name: projectId
        schema:
          type: integer
        description: A unique integer value identifying this project.
        required: true
      tags:
      - Projects
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSubscriptionState'
          description: ''
        '204':
          description: Subscription updated
...
components:
  schemas:
...
    ProjectMembership:
      type: object
      properties:
        projectMembershipId:
          type: integer
          readOnly: true
        projectId:
          type: integer
          readOnly: true
        userId:
          type: integer
          readOnly: true
        username:
          type: string
          readOnly: true
        firstName:
          type: string
          readOnly: true
        lastName:
          type: string
          readOnly: true
        email:
          type: string
          format: email
          readOnly: true
        role:
          type: string
          readOnly: true
        profileImg:
          type: string
          readOnly: true
      required:
      - email
      - firstName
      - lastName
      - profileImg
      - projectId
      - projectMembershipId
      - role
      - userId
      - username
...
    ProjectSubscriptionState:
      type: object
      properties:
        subscribed:
          type: boolean
      required:
      - subscribed
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRFToken